requirements.txt
face_recognition==1.3.0
numpy==1.24.0
opencv_python==4.6.0.66
pandas==1.4.4
Pillow==9.3.0
streamlit==1.16.0
streamlit_option_menu==0.3.2
